django/requirements.txt
django-environ
gunicorn
umap-project
